Convert remaining objective c code to swift

+++ TradeItIosTicketSDK2/TradeItFxOrderCapabilities.swift
@@ -1,3 +1,3 @@
 class TradeItFxOrderCapabilities: TradeItInstrumentOrderCapabilities {
-    var leverageOptions: [Double]?
+    var leverageOptions: [Int]?
 }
